react-native build-desktop command added
