+++ local-cli/desktop/buildDesktop.js
@@ -15,128 +15,18 @@
 
 const path = require('path');
 const chalk = require('chalk');
-const fs = require('fs');
 const child_process = require('child_process');
+const parseArguments = require('./parseArguments');
+const build = require('./build');
 
 
-function buildDesktop(args, dependencies) {
-  var desktopExternalModules = _findDesktopExternalModules(args);
-  console.log("Found desktop external modules: " + desktopExternalModules);
-  var desktopJSBundlePath = _findDesktopJSBundlePath(args);
-  console.log("Found desktop JS bundle path: " + desktopJSBundlePath);
-  return _findModules(args).then((dependencies) => {
-    return new Promise((resolve, reject) => {
-      _findDesktopModules(args, dependencies, resolve, reject);
-    });
-  }).then((desktopModules) => {
-    return new Promise((resolve, reject) => {
-      if (desktopModules && desktopModules.length > 0) {
-        _buildModules(args, desktopModules, resolve, reject);
-      } else {
-        resolve();
-      }
-    });
-  }).then(() => {
-    return _buildApplication(args, desktopExternalModules, desktopJSBundlePath);
-  });
+function buildDesktop(argv, config) {
+  const args = parseArguments(argv);
+  build(args);
 }
 
-function _findDesktopExternalModules(args) {
-  var data = fs.readFileSync(path.join(args.root, 'package.json'));
-  return JSON.parse(data).desktopExternalModules;
-}
-
-function _findDesktopJSBundlePath(args) {
-  var data = fs.readFileSync(path.join(args.root, 'package.json'));
-  return JSON.parse(data).desktopJSBundlePath;
-}
-
-function _findModules(args) {
-  return new Promise((resolve, reject) => {
-    fs.readFile(path.join(args.root, 'package.json'), (err, data) => {
-      if (err) {
-        reject(err);
-      } else {
-        resolve(JSON.parse(data).dependencies);
-      }
-    });
-  });
-}
-
-function _findDesktopModules(args, dependencies, resolve, reject) {
-  var potential = Object.keys(dependencies).map((p) => {
-    if (p === 'react' || p === 'react-native')
-      return null;
-
-    return new Promise((resolve, reject) => {
-      const depPath = path.join(args.root, 'node_modules', p);
-      fs.readFile(path.join(depPath, 'package.json'), (err, data) => {
-        const desktop = data && JSON.parse(data)._desktop;
-        if (desktop !== undefined && desktop.hasOwnProperty('build')) {
-          resolve({name: p, build: desktop.build, path: depPath});
-        } else {
-          resolve(null);
-        }
-      });
-    });
-  });
-
-  Promise.all(potential).then((result) => {
-    resolve(result.filter((v) => v !== null));
-  }).catch((err) => {
-    reject("Error assessing Desktop module: " + err);
-  });
-}
-
-function _buildModules(args, dependencies, resolve, reject) {
-  var builds = dependencies.map((p) => {
-    return new Promise((resolve, reject) => {
-      console.log(chalk.bold('Building Desktop module: ') + p.name);
-
-      var buildCommand = 'URN_OUTPUT_DIR=' + path.resolve(args.root, 'desktop', 'plugins');
-      buildCommand += ' ' + p.build;
-
-      child_process.exec(buildCommand, {cwd: p.path}, (error, stdout, stderr) => {
-                             if (error) {
-                               reject(error);
-                             } else {
-                               resolve(true);
-                             }
-                            });
-    });
-  });
-
-  Promise.all(builds).then((result) => {
-    resolve();
-  }).catch((err) => {
-    reject("Package build failed: " + err);
-  });
-}
-
-function _buildApplication(args, desktopExternalModules, desktopJSBundlePath) {
-  return new Promise((resolve, reject) => {
-    console.log(chalk.bold('Building the app...'));
-
-    var buildCommand = process.platform === "win32" ? "build.bat" : "./build.sh";
-    if (typeof desktopExternalModules !== 'undefined' && desktopExternalModules !== null) {
-      buildCommand += ' -e "' + desktopExternalModules.toString().replace(/,/g, ';') + '"';
-    }
-    if (typeof desktopJSBundlePath !== 'undefined' && desktopJSBundlePath !== null) {
-      buildCommand += ' -j "' + desktopJSBundlePath.toString() + '"';
-    }
-    if (process.platform === "win32") {
-      buildCommand += ' -g "' + "MinGW Makefiles" + '"';
-    }
-    child_process.exec(buildCommand, {cwd: path.join(args.root, 'desktop')},
-                        (error, stdout, stderr) => {
-                          if (error)
-                            reject(error);
-                          else {
-                            console.log(stdout);
-                            resolve();
-                          }
-                        });
-  });
-}
-
-module.exports = buildDesktop;
+module.exports = {
+  name: 'build-desktop',
+  description: 'builds your app as native application',
+  func: buildDesktop
+};
